requirements.txt
-f https://cdn.research.pdx.edu/pypi/dist/
varlet
setuptools
psycopg2
django-cloak
pytz
django-arcutils>=1.2.0,<2.0
ldap3
mock
model-mommy
coverage
flake8
django<1.8
